d_a_npc_so: missing-ctor class FIXED — inline daNpc_So_c(){} in header emits __ct__10daNpc_So_cFv WEAK (retail scope:weak, 0x3F38 size 0x244) at 100.0 per-symbol + materializes the full member-ctor weak chain (mObjAcch/mAcchCir/mStts/mCyl/mSph/mBtpAnm/mRipple vtables now present); D5 precondition DEBUNKED: 0x6F4 is the cCcD_Stts vtable slot INSIDE mStts2@0x6DC (header correct, dCcD_Stts = 0x18 header + GStts@0x6F8), not a separate object; fresh .o postdates

+++ src/d/actor/d_a_npc_so.cpp
@@ -412,7 +412,7 @@ u16 daNpc_So_c::next_msgStatus(u32* pMsgNo) {
         return msg_status;
     }
 
-    switch (*pMsgNo) {
+    switch ((s32)*pMsgNo) {
     case 0x32CA:
         *pMsgNo = 0x32CB;
         break;
